+++ R/ping-package.r
@@ -1,10 +1,8 @@
 
-#' Check if a remote computer is up
+#' Check if the local or remote computer is up
 #'
 #' @useDynLib pingr, .registration = TRUE
-#' @docType package
-#' @name pingr-package
-NULL
+"_PACKAGE"
 
 #' Check if a port of a server is active, measure response time
 #'
@@ -124,39 +122,44 @@ ping_os <- function(destination, continuous, count, timeout) {
   list(cmd = cmd, regex = "^.*time=(.+)[ ]?ms.*$")
 }
 
-## Domains and IPs to test for internet connection
-
-internet_domains <- c("google-public-dns-a.google.com",
-                      "b.resolvers.Level3.net")
-
-internet_ips <- c("8.8.8.8",
-                  "4.2.2.2")
-
 #' Is the computer online?
 #'
-#' Ping some name servers that are always (well, almost) up.
-#' If these are unreachable, then you are most probably not online.
+#' Check if the computer is online. It does three tries:
+#' * Queries myip.opendns.com on OpenDNS, see [my_ip()].
+#' * Retrieves icanhazip.com via HTTPS, see [my_ip()].
+#' * Retrieve Apple's Captive Portal test page, see [apple_captive_test()].
+#' If any of these are successful, it returns `TRUE`.
 #'
-#' @param timeout Timeout for the pings.
+#' @param timeout Timeout for the queries. (Note: it is currently not
+#'   used for the DNS query.)
 #' @return Possible values: \itemize{
 #'   \item \code{TRUE} Yes, online.
 #'   \item \code{FALSE} No, not online.
-#'   \item \code{"nodns"} We re online, but without a DNS service.
 #' }
 #'
 #' @export
 #' @examples
 #' is_online()
-#' is_online(timeout = 0.01)
 
-is_online <- function(timeout = 0.2) {
-  for (domain in internet_domains) {
-    if (!is.na(ping(domain, count = 1, timeout = timeout))) { return(TRUE) }
-  }
-  for (ip in internet_ips) {
-    if (!is.na(ping(ip, count = 1, timeout = timeout))) { return("nodns") }
-  }
-  return(FALSE)
+is_online <- function(timeout = 1) {
+  opts <- options(timeout = timeout)
+  on.exit(options(opts), add = TRUE)
+
+  tryCatch({
+    my_ip(method = "dns")
+    return(TRUE)
+  }, error = function(e) NULL)
+
+  tryCatch({
+    my_ip(method = "https")
+    return(TRUE)
+  }, error = function(e) NULL)
+
+  tryCatch({
+    if (apple_captive_test()) return(TRUE)
+  }, error = function(e) NULL)
+
+  FALSE
 }
 
 #' `is_up()` checks if a web server is up.