+++ NEWS.md
@@ -1,4 +1,19 @@
 
+# 2.0.0
+
+* New `nsl()` function to perform DNS queries.
+
+* New `my_ip()` function to query the computer's public IP address.
+
+* New `apple_captive_test()` function to check Apple's captive test
+  web page to see if the computer is online.
+
+* Better `is_online()` implementation, it uses DNS and HTTPS instead
+  of an ICMP ping via an external ping program.
+
+* Now `ip_up()` checks first if the computer is connected to the
+  internet, via `is_online()`.
+
 # 1.2.0
 
 * New `is_up()` function to check if a web (other other TCP) server is up.